requirements.txt
chromadb
langchain
langchain_experimental
langchainhub
langchain-cli
langchain_openai
python-dotenv
feedparser
newspaper3k
listparser
langserve[all]
